fix: merge shufflingDecisionRoot.ts to shuffling.ts
